RavenDB-26136 : refactor PowerBIFetchQuery - move all the regex-based path to a separate class (still use it as fallback)
